refactor: so only one interface for integration provider exists
