+++ apps/web/src/pages/integrations/IntegrationsStoreModal.tsx
@@ -3,11 +3,8 @@ import styled from '@emotion/styled';
 import { Grid, Group, Modal, ActionIcon, createStyles, MantineTheme, Drawer } from '@mantine/core';
 import {
   ChannelTypeEnum,
-  IConfigCredentials,
-  ILogoFileName,
   EmailProviderIdEnum,
   InAppProviderIdEnum,
-  ProvidersIdEnum,
   SmsProviderIdEnum,
   ICredentials,
 } from '@novu/shared';
@@ -25,6 +22,7 @@ import { useSegment } from '../../components/providers/SegmentProvider';
 import { IntegrationsStoreModalAnalytics } from './constants';
 import { NovuSmsProviderModal } from './components/NovuSmsProviderModal';
 import { useCreateInAppIntegration } from '../../hooks/useCreateInAppIntegration';
+import { IIntegratedProvider } from './IntegrationsStorePage';
 
 export function IntegrationsStoreModal({
   scrollTo,
@@ -313,22 +311,6 @@ const useDrawerStyles = createStyles((theme: MantineTheme) => {
     },
   };
 });
-
-export interface IIntegratedProvider {
-  providerId: ProvidersIdEnum;
-  integrationId: string;
-  displayName: string;
-  channel: ChannelTypeEnum;
-  credentials: IConfigCredentials[];
-  docReference: string;
-  comingSoon: boolean;
-  active: boolean;
-  connected: boolean;
-  logoFileName: ILogoFileName;
-  betaVersion: boolean;
-  novu?: boolean;
-  environmentId?: string;
-}
 
 export interface IntegrationEntity {
   _id?: string;
