+++ apps/web/src/pages/templates/components/ProvidersPage.tsx
@@ -2,7 +2,8 @@ import { Center, Loader } from '@mantine/core';
 import { ChannelTypeEnum } from '@novu/shared';
 import { useState } from 'react';
 import { colors } from '../../../design-system';
-import { IIntegratedProvider, IntegrationsStoreModal } from '../../integrations/IntegrationsStoreModal';
+import { IntegrationsStoreModal } from '../../integrations/IntegrationsStoreModal';
+import { IIntegratedProvider } from '../../integrations/IntegrationsStorePage';
 import { useProviders } from '../../integrations/useProviders';
 import { ListProviders } from './ListProviders';
 import { SubPageWrapper } from './SubPageWrapper';
